feat: Updated query for rides
