+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bcc",
-  "version": "1.14.2",
+  "version": "1.14.3",
   "private": true,
   "scripts": {
     "dev": "next dev",
