+++ src/components/RideDetails/RideInfo.tsx
@@ -7,7 +7,7 @@ type RowProps = {
 };
 
 const Row = ({ children }: RowProps) => (
-  <div className="grid w-full grid-cols-[100px_1fr] items-center justify-between px-2 font-medium md:grid-cols-[220px_1fr] md:justify-start md:gap-4">
+  <div className="grid w-full grid-cols-[100px_1fr] items-center justify-between px-2 font-medium md:grid-cols-[220px_1fr] md:justify-start md:gap-4 gap-2">
     {children}
   </div>
 );
@@ -114,13 +114,13 @@ export const RideInfo = ({ ride }: Props) => {
       )}
 
       {(riderNotes || []).length > 0 && (
-        <div className="flex w-full flex-col gap-2 rounded bg-white py-2 shadow-md">
+        <div className="flex w-full flex-col rounded bg-white py-2 shadow-md divide-y divide-neutral-100">
           <div className="px-2 text-xl font-bold tracking-wide text-neutral-700">
             Messages
           </div>
           {riderNotes?.map((rider) => (
             <Row key={rider.name}>
-              <div className="whitespace-pre-line">{rider.name}</div>
+              <div className="whitespace-pre-line truncate">{rider.name}</div>
               <div className="whitespace-pre-line">{rider.rideNotes}</div>
             </Row>
           ))}
